#54 Apply deltas to address. Bitcoin. Fix structure.
